+++ src/Jarboe/Table/Fields/Traits/Ajax.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Yaro\Jarboe\Table\Fields\Traits;
+
+trait Ajax
+{
+    protected $ajax = false;
+
+    public function ajax(bool $ajax = true)
+    {
+        $this->ajax = $ajax;
+
+        return $this;
+    }
+
+    public function isAjax()
+    {
+        return $this->ajax;
+    }
+}