+++ src/resources/views/crud/fields/tags/edit.blade.php
@@ -1,7 +1,13 @@
+<?php
+/** @var \Yaro\Jarboe\Table\Fields\Tags $field */
+?>
 <label class="label">{{ $field->title() }}</label>
 <label class="input {{ $errors->has($field->name()) ? 'state-error' : '' }}">
-
-    <select class="select-2--tags form-control" multiple="multiple" name="{{ $field->name() }}[]">
+    <select class="select-2--tags form-control"
+            multiple="multiple"
+            data-relation-search-url="{{ $field->isAjax() ? $field->getRelationSearchUrl() : '' }}"
+            data-original-name="{{ $field->name() }}"
+            name="{{ $field->name() }}[]">
         @if ($field->hasOld())
             @foreach (($field->old() ?: []) as $value)
                 <option value="{{ $value }}" selected>{{ $value }}</option>
@@ -13,7 +19,7 @@
                 @endforeach
             @endif
         @else
-            @if ($field->isOptionsHidden())
+            @if ($field->isOptionsHidden() || $field->isAjax())
                 @foreach ($field->getSelectedOptions($model) as $id => $value)
                     <option value="{{ $value }}" selected>{{ $value }}</option>
                 @endforeach
@@ -24,7 +30,6 @@
             @endif
         @endif
     </select>
-
 </label>